refactor(src) phpstan l0 fixes
