+++ src/Codeception/Module/WPBrowser.php
@@ -17,48 +17,8 @@ class WPBrowser extends PhpBrowser
      * @var array
      */
     protected $requiredFields = ['adminUsername', 'adminPassword', 'adminPath'];
-    /**
-     * The login screen absolute URL
-     *
-     * @var string
-     */
-    protected $loginUrl;
-    /**
-     * The admin absolute URL.
-     *
-     * @var [type]
-     */
-    protected $adminPath;
-    /**
-     * The plugin screen absolute URL
-     *
-     * @var string
-     */
-    protected $pluginsPath;
 
-    /**
-     * Initializes the module setting the properties values.
-     */
-    public function _initialize()
-    {
-        parent::_initialize();
-
-        $this->configBackCompat();
-
-        $adminPath = $this->config['adminPath'];
-        $this->loginUrl = str_replace('wp-admin', 'wp-login.php', $adminPath);
-        $this->adminPath = rtrim($adminPath, '/');
-        $this->pluginsPath = $this->adminPath . '/plugins.php';
-    }
-
-    protected function configBackCompat()
-    {
-        if (isset($this->config['adminUrl']) && !isset($this->config['adminPath'])) {
-            $this->config['adminPath'] = $this->config['adminUrl'];
-        }
-    }
-
-    /**
+	/**
      * Returns all the cookies whose name matches a regex pattern.
      *
      * @example
